feat: use Box component, update routes.ts
